refactor(paging): the shared page carries a plain int, and two more caps land
The lint caught five int -> int32 narrowings in the page adapters -- offsets no
Bitbucket will ever return, each one a suppression waiting to be written. The
shared Page now carries NextPageStart as an int, so every adapter widens and
none narrows: openapi.Offset lifts the generated client's *int32, and the
endpoints decoded by hand pass their own counted offset straight through.

Two caps that were renamed last commit but not wired: reposettings' permission
listings took maxResults and then handed a constant to the walk, so the value
was ineffectual -- which is exactly what the linter said. They cap now, and the
package exports AllResults for the callers that want everything.

Those callers are the dry runs. `permission grant --dry-run` read the first 100
entries to decide create-or-update, which was harmless while the listing ignored
the number and is not harmless now: a project or repository past 100 entries
would predict "create" for a subject that already has a permission. Both grant
and revoke previews ask for all of them, the same trap AllResults was added for
in #470.

Also drops a second orphaned test helper the linter found, and a conversion in a
live test that stopped being a conversion when the comment state's type changed.

+++ internal/services/browse/service.go
@@ -96,7 +96,7 @@ func (service *Service) Tree(ctx context.Context, repo RepositoryRef, path strin
 			isLastPage := response.IsLastPage
 			page := openapi.Page[string]{Values: names, IsLastPage: &isLastPage}
 			if response.NextPageStart != nil {
-				next := int32(*response.NextPageStart)
+				next := *response.NextPageStart
 				page.NextPageStart = &next
 			}
